machine_learning/logistic_regression/char_reco/requirements.txt
streamlit
numpy
scikit-learn
opencv-python
Pillow
streamlit-drawable-canvas
